web/just_wasting_time/requirements.txt
flask
pyjwt
bcrypt
sqlalchemy
python-dotenv
flask_sqlalchemy